requirements.txt
django
django-environ
gunicorn
python-dateutil
requests
youtube-dl